+++ src/melee/ty/tylist.c
@@ -44,7 +44,7 @@
                                            float x, float y, float z);
 /* 313774 */ static void _tyList_80313774(void);
 /* 31438C */ static void _tyList_8031438C(HSD_GObj* arg0);
-/* 314504 */ static void _tyList_80314504(HSD_GObj* gobj, int);
+/* 314504 */ static void _tyList_80314504(HSD_GObj* gobj, intptr_t);
 /* 31457C */ static void _tyList_8031457C(void);
 /* 3148E4 */ static void _tyList_803148E4(s32 arg0);
 /* 4A2AC0 */ static TyListState _tyList_804A2AC0;
@@ -886,7 +886,7 @@ void _tyList_8031438C(HSD_GObj* gobj)
     HSD_GObj_80390CD4(entry->x0);
 }
 
-void _tyList_80314504(HSD_GObj* gobj, int unused)
+void _tyList_80314504(HSD_GObj* gobj, intptr_t unused)
 {
     TyListData* data = (TyListData*) gobj;
